Rename typing module to avoid clash with built-in python module
