+++ pyprobe/analysis/degradation_mode_analysis.py
@@ -15,8 +15,8 @@ from scipy.interpolate import PPoly
 import pyprobe.analysis.base.degradation_mode_analysis_functions as dma_functions
 from pyprobe.analysis import smoothing, utils
 from pyprobe.analysis.utils import AnalysisValidator
+from pyprobe.pyprobe_types import FilterToCycleType, PyProBEDataType
 from pyprobe.result import Result
-from pyprobe.typing import FilterToCycleType, PyProBEDataType
 
 
 def _get_gradient(
