+++ pyprobe/analysis/differentiation.py
@@ -9,8 +9,8 @@ from pydantic import BaseModel, validate_call
 
 import pyprobe.analysis.base.differentiation_functions as diff_functions
 from pyprobe.analysis.utils import AnalysisValidator
+from pyprobe.pyprobe_types import PyProBEDataType
 from pyprobe.result import Result
-from pyprobe.typing import PyProBEDataType
 
 
 @validate_call
